+++ core/src/sound_system/mod.rs
@@ -1,46 +1,14 @@
 use crate::wgl;
 use pest::error::Error;
-
-use rand::prelude::*;
-use std::str::FromStr;
-
 mod compiler;
 mod distribution;
+mod generator;
 mod types;
 
 pub mod phone;
 pub mod rules;
 pub use compiler::*;
 pub use types::*;
-
-impl MonoSyllableRepartition {
-    pub fn into_percentage(self) -> f32 {
-        match self {
-            MonoSyllableRepartition::Always => 1.0,
-            MonoSyllableRepartition::Mostly => 0.85,
-            MonoSyllableRepartition::Frequent => 0.5,
-            MonoSyllableRepartition::LessFrequent => 0.20,
-            MonoSyllableRepartition::Rare => 0.07,
-            MonoSyllableRepartition::Never => 0.0,
-        }
-    }
-}
-
-impl FromStr for MonoSyllableRepartition {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "always" => Ok(MonoSyllableRepartition::Always),
-            "mostly" => Ok(MonoSyllableRepartition::Mostly),
-            "frequent" => Ok(MonoSyllableRepartition::Frequent),
-            "less_frequent" => Ok(MonoSyllableRepartition::LessFrequent),
-            "rare" => Ok(MonoSyllableRepartition::Rare),
-            "never" => Ok(MonoSyllableRepartition::Never),
-            _ => Err("no match"),
-        }
-    }
-}
 
 pub fn from_string(input: &'_ str) -> Result<SoundSystem, Error<wgl::Rule>> {
     wgl::from_string(input).map(SoundSystem::compile)
@@ -63,48 +31,21 @@ impl SoundSystem {
         self.sort_phonemes()
     }
 
-    pub fn generate_words(
-        &self,
-        number: usize,
-        repartition: MonoSyllableRepartition,
-    ) -> Vec<String> {
-        let mut words = vec![];
-        let percentage = repartition.into_percentage();
-        for _ in 0..number {
-            let mut number_of_syllables = 1;
-            if percentage < 1.0 && random::<f32>() > percentage {
-                number_of_syllables += 1 + distribution::power_law(4, 0.5);
-            }
-            let mut word = String::new();
-            for _ in 0..number_of_syllables {
-                word.push_str(&self.syllable());
-            }
-            if !words.contains(&word) {
-                words.push(word);
-            }
-        }
-        words
-    }
+    pub fn phonology(&self) -> phone::Phones {
+        let mut vec: phone::Phones = self
+            .phonemes_sorted()
+            .clone()
+            .into_iter()
+            .flat_map(|(_, (phones, _))| phones)
+            .filter(|phone| match phone {
+                phone::Phone::Diacritic(_) => false,
+                _ => true,
+            })
+            .collect();
+        vec.sort_unstable_by(|left_phones, right_phones| Ord::cmp(&right_phones, &left_phones));
 
-    fn syllable(&self) -> String {
-        let syllables_size = self.syllables().len();
-        let syllable_drop = syllable_drop(syllables_size);
-        let mut syllable = String::new();
-        let index = distribution::power_law(syllables_size, syllable_drop);
-        let pattern = &self.syllables()[index];
-
-        for class_name in pattern {
-            if let Some(letters) = self.classes().get(class_name) {
-                let distribution = self
-                    .distribution()
-                    .iter()
-                    .filter(|t| letters.contains(&t.0))
-                    .collect::<Vec<_>>();
-                let letter = distribution::select(distribution);
-                syllable.push_str(letter);
-            }
-        }
-        syllable
+        vec.dedup();
+        vec
     }
 
     pub fn ipa_representation(&self, word: &'_ str) -> String {
@@ -167,10 +108,14 @@ impl SoundSystem {
             Condition::Not(cond_type) => {
                 !self.resolve_condition_type(input, letter, position, length, cond_type)
             }
-            Condition::Binary(op, left, right) => {
+            Condition::Binary {
+                operand,
+                left,
+                right,
+            } => {
                 let left_bool = self.resolve_condition(input, letter, position, length, left);
                 let right_bool = self.resolve_condition(input, letter, position, length, right);
-                match op {
+                match operand {
                     ConditionOperand::And => left_bool && right_bool,
                     ConditionOperand::Or => left_bool || right_bool,
                 }
@@ -217,13 +162,5 @@ impl SoundSystem {
             }
             _ => false,
         }
-    }
-}
-
-fn syllable_drop(number_of_syllables: usize) -> f32 {
-    if number_of_syllables < 9 {
-        0.6 - (number_of_syllables as f32) * 0.05
-    } else {
-        0.12
     }
 }