+++ core/src/sound_system/phone/types.rs
@@ -1,5 +1,5 @@
 pub type Phones = Vec<Phone>;
-
+use std::fmt;
 #[derive(Debug, Eq, PartialEq, Serialize, Clone, Copy)]
 pub enum Phone {
     Consonant(Consonant),
@@ -54,31 +54,65 @@ pub struct Consonant {
 pub enum ConsonantPlace {
     Bilabial,
     LabioDental,
-    LabioVelar,
     Dental,
     Alveolar,
     PostAlveolar,
     Retroflex,
     Palatal,
     Velar,
+    LabioVelar,
     Uvular,
     Pharyngeal,
-    Glotal,
+    Glottal,
 }
 
 #[derive(Debug, Eq, PartialEq, Serialize, Clone, Ord, PartialOrd, Copy)]
 #[repr(u8)]
 pub enum ConsonantManner {
+    Nasal,
     Stop,
     SibilantFricative,
     Fricative,
-    Nasal,
     Trill,
     Tap,
     LateralFricative,
     Approximant,
     LateralApproximant,
     LateralTap,
+}
+
+impl fmt::Display for ConsonantManner {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let s = match self {
+            ConsonantManner::Nasal => "Nasal",
+            ConsonantManner::Stop => "Stop",
+            ConsonantManner::SibilantFricative => "Sibilant fricative",
+            ConsonantManner::Fricative => "Fricative",
+            ConsonantManner::Trill => "Trill",
+            ConsonantManner::Tap => "Tap",
+            ConsonantManner::LateralFricative => "Lateral fricative",
+            ConsonantManner::Approximant => "Approximant",
+            ConsonantManner::LateralApproximant => "Lateral approximant",
+            ConsonantManner::LateralTap => "Lateral tap",
+        };
+        write!(f, "{}", s)
+    }
+}
+
+impl fmt::Display for Height {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        use Height::*;
+        let s = match self {
+            Close => "Close",
+            NearClose => "Near close",
+            CloseMid => "Close mid",
+            Mid => "Mid",
+            OpenMid => "Open mid",
+            NearOpen => "Near open",
+            Open => "Open",
+        };
+        write!(f, "{}", s)
+    }
 }
 
 #[derive(Debug, Eq, PartialEq, Serialize, Clone, Ord, PartialOrd, Copy)]
@@ -176,15 +210,4 @@ pub enum Tone {
     Falling,
     Bottom,
     Downstep,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::cmp::Ordering;
-    #[test]
-    fn test_manner_order() {
-        let result = ConsonantManner::Nasal.cmp(&ConsonantManner::Stop);
-        assert_eq!(result, Ordering::Greater)
-    }
 }